committed changes

+++ core/src/main/java/io/eliteblue/erp/core/bean/ErpDetachmentForm.java
@@ -3,10 +3,10 @@ package io.eliteblue.erp.core.bean;
 import io.eliteblue.erp.core.lazy.LazyEmployeeModel;
 import io.eliteblue.erp.core.lazy.LazyErpPostModel;
 import io.eliteblue.erp.core.model.*;
-import io.eliteblue.erp.core.service.ErpClientService;
-import io.eliteblue.erp.core.service.ErpDetachmentService;
-import io.eliteblue.erp.core.service.ErpEmployeeService;
-import io.eliteblue.erp.core.service.OperationsAreaService;
+import io.eliteblue.erp.core.service.*;
+import io.eliteblue.erp.core.util.CurrentUserUtil;
+import io.eliteblue.erp.core.util.ExcelUtils;
+import org.apache.poi.ss.usermodel.Workbook;
 import org.omnifaces.util.Faces;
 import org.primefaces.event.SelectEvent;
 import org.primefaces.event.UnselectEvent;
@@ -15,14 +15,21 @@ import org.primefaces.util.LangUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.faces.application.FacesMessage;
+import javax.faces.context.ExternalContext;
 import javax.faces.context.FacesContext;
 import javax.faces.model.SelectItem;
 import javax.faces.view.ViewScoped;
 import javax.inject.Named;
+import javax.servlet.ServletOutputStream;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.io.Serializable;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
+import java.util.stream.Collectors;
 
 import static com.github.adminfaces.template.util.Assert.has;
 
@@ -31,6 +38,7 @@ import static com.github.adminfaces.template.util.Assert.has;
 public class ErpDetachmentForm implements Serializable {
 
     final DateFormat format = new SimpleDateFormat("HH:mm");
+    private final DateFormat format2 = new SimpleDateFormat("MM/dd/yyyy");
 
     @Autowired
     private ErpDetachmentService erpDetachmentService;
@@ -43,6 +51,15 @@ public class ErpDetachmentForm implements Serializable {
 
     @Autowired
     private ErpEmployeeService erpEmployeeService;
+
+    @Autowired
+    private ErpEquipmentService erpEquipmentService;
+
+    @Autowired
+    private ErpSilService silService;
+
+    @Autowired
+    private CurrentUserUtil currentUserUtil;
 
     private Long id;
     private Long clientId;
@@ -64,6 +81,7 @@ public class ErpDetachmentForm implements Serializable {
     private List<ErpEmployee> filteredEmployees;
     private LazyDataModel<ErpEmployee> lazyErpEmployees;
     private ErpEmployee selectedEmployee;
+    private Map<String, Boolean> exessOTSelect;
 
     public void init() {
         if(Faces.isAjaxRequest()) {
@@ -94,6 +112,9 @@ public class ErpDetachmentForm implements Serializable {
             SelectItem itm = new SelectItem(o.getLocation(), o.getLocation());
             locations.add(itm);
         }
+        exessOTSelect = new HashMap<>();
+        exessOTSelect.put("NO", false);
+        exessOTSelect.put("YES", true);
     }
 
     public Long getId() {
@@ -232,6 +253,14 @@ public class ErpDetachmentForm implements Serializable {
         this.lazyErpEmployees = lazyErpEmployees;
     }
 
+    public Map<String, Boolean> getExessOTSelect() {
+        return exessOTSelect;
+    }
+
+    public void setExessOTSelect(Map<String, Boolean> exessOTSelect) {
+        this.exessOTSelect = exessOTSelect;
+    }
+
     public DateFormat getFormat() {
         return format;
     }
@@ -327,9 +356,16 @@ public class ErpDetachmentForm implements Serializable {
     public void remove() throws Exception {
         if(has(erpDetachment) && has(erpDetachment.getId())) {
             String name = erpDetachment.getName();
-            erpDetachmentService.delete(erpDetachment);
-            addDetailMessage("DETACHMENT DELETED", name, FacesMessage.SEVERITY_INFO);
-            FacesContext.getCurrentInstance().getExternalContext().redirect("client-form.xhtml?id="+clientId);
+            List<ErpDetachment> dets = new ArrayList<>();
+            dets.add(erpDetachment);
+            List<ErpEquipment> erpEquipments = erpEquipmentService.getAllDetachmentIn(dets);
+            if(erpEquipments == null || erpEquipments.size() == 0) {
+                erpDetachmentService.delete(erpDetachment);
+                addDetailMessage("DETACHMENT DELETED", name, FacesMessage.SEVERITY_INFO);
+                FacesContext.getCurrentInstance().getExternalContext().redirect("client-form.xhtml?id=" + clientId);
+            } else {
+                addDetailMessage("DELETE DETACHMENT FAILED", " Equipments are added in this detachment. Delete them first.", FacesMessage.SEVERITY_ERROR);
+            }
         }
     }
 
@@ -342,6 +378,147 @@ public class ErpDetachmentForm implements Serializable {
             addDetailMessage("DETACHMENT SAVED", erpDetachment.getName(), FacesMessage.SEVERITY_INFO);
             FacesContext.getCurrentInstance().getExternalContext().redirect("client-form.xhtml?id="+clientId);
         }
+    }
+
+    public void downloadSILReport() throws Exception {
+        LocalDate today = LocalDate.now();
+        final Double maxSIL = 5.0;
+        if (has(erpDetachment)) {
+            ExcelUtils.initializeWithFilename("sil_template.xlsx", "Sheet1");
+            System.out.println("Excel Initialization DONE...");
+            String detachmentName = erpDetachment.getName().replaceAll(" ", "_");
+
+            ExcelUtils.setCell(3,2, capitalizeString(erpDetachment.getName().toLowerCase()));
+            if(has(erpDetachment.getAssignedEmployees())) {
+                ExcelUtils.setCell(3, 4, erpDetachment.getAssignedEmployees().size());
+                int noCnt = 1;
+                int x = 8;
+                List<ErpEmployee> employees = new ArrayList<>(erpDetachment.getAssignedEmployees());
+                Collections.sort(employees, (o1, o2) -> (o1.getLastname().compareTo(o2.getLastname())));
+                for(ErpEmployee e: employees) {
+                    // get SIL by employee
+                    List<ErpSil> sils = silService.findAllByEmployee(e);
+                    for(ErpSil s: sils) {
+                        ExcelUtils.setCell(x, 0, noCnt);
+                        ExcelUtils.setCell(x, 1, e.getStatus().name());
+                        ExcelUtils.setCell(x, 2, detachmentName.replaceAll("_", " "));
+                        String empFullName = e.getLastname() + ", " + e.getFirstname();
+                        ExcelUtils.setCell(x, 3, empFullName);
+                        if (has(e.getJoinedDate())) {
+                            ExcelUtils.setCell(x, 4, format2.format(e.getJoinedDate()));
+                        }
+                        if (has(e.getResignedDate())) {
+                            ExcelUtils.setCell(x, 5, format2.format(e.getResignedDate()));
+                        }
+                        ExcelUtils.setCell(x, 6, s.getNoDaysAvailed());
+                        ExcelUtils.setCell(x, 7, s.getAmountAvailed());
+                        SimpleDateFormat sd1 = new SimpleDateFormat("MMM dd");
+                        SimpleDateFormat sd2 = new SimpleDateFormat("dd, yyyy");
+                        SimpleDateFormat sd3 = new SimpleDateFormat("MMM dd yyyy");
+                        //String availedDate = sd1.format(s.getDateAvailedStart()) + sd2.format(s.getDateAvailedStop());
+                        String availedDate = Optional.ofNullable(s.getDateAvailedStart())
+                                .map(sd1::format)
+                                .orElse("") + " - " + Optional.ofNullable(s.getDateAvailedStop())
+                                .map(sd2::format)
+                                .orElse("");
+                        ExcelUtils.setCell(x, 8, availedDate);
+                        //String paymentDate = sd1.format(s.getDateOfPaymentStart()) + sd2.format(s.getDateOfPaymentStop());
+                        String paymentDate = Optional.ofNullable(s.getDateOfPaymentStart())
+                                .map(sd1::format)
+                                .orElse("") + " - " + Optional.ofNullable(s.getDateOfPaymentStop())
+                                .map(sd2::format)
+                                .orElse("");
+                        ExcelUtils.setCell(x, 9, paymentDate);
+                        ExcelUtils.setCell(x, 10, s.getModeOfPayment().name());
+                        ExcelUtils.setCell(x, 11, s.getNoDaysUnAvailed());
+                        ExcelUtils.setCell(x, 12, s.getAmountUnAvailed());
+                        //String coveredDate = sd3.format(s.getDateCoveredStart()) + " - " + sd3.format(s.getDateCoveredStop());
+                        String coveredDate = Optional.ofNullable(s.getDateCoveredStart())
+                                .map(sd3::format)
+                                .orElse("") + " - " + Optional.ofNullable(s.getDateCoveredStop())
+                                .map(sd3::format)
+                                .orElse("");
+                        ExcelUtils.setCell(x, 13, coveredDate);
+                        noCnt++;
+                        x++;
+                    }
+                }
+                ExcelUtils.evaluateAllCells();
+            }
+
+            generateFile(ExcelUtils.workbook, "SIL_REPORT_"+detachmentName+today.format(DateTimeFormatter.ofPattern("MMddyyyy")));
+            System.out.println("Generation SUCCESS!!");
+            ExcelUtils.workbook.close();
+        }
+    }
+
+    public void downloadFile() throws Exception {
+        LocalDate today = LocalDate.now();
+        String detachmentName = erpDetachment.getName().replaceAll(" ", "_");
+        if (has(erpDetachment)) {
+            ExcelUtils.initializeWithFilename("assigned_emp.xlsx", "Sheet1");
+            System.out.println("Excel Initialization DONE...");
+            boolean hasAssignedEmployees = (has(erpDetachment) && erpDetachment.getAssignedEmployees() != null && erpDetachment.getAssignedEmployees().size() > 0);
+            ExcelUtils.setCell(1, 1, format2.format(new Date()));
+            ExcelUtils.setCell(2,1, capitalizeString(erpDetachment.getName().toUpperCase()));
+            String currentUserName = CurrentUserUtil.getFullName();
+            ExcelUtils.setCell(3, 1, currentUserName);
+            if(hasAssignedEmployees) {
+                int x = 7;
+                List<ErpEmployee> emps = new ArrayList<>();
+                emps.addAll(erpDetachment.getAssignedEmployees());
+                emps.sort(Comparator.comparing((ErpEmployee ewa) -> ewa.getLastname()));
+                for(ErpEmployee emp : emps) {
+                    ExcelUtils.setCell(x, 0, emp.getEmployeeId());
+                    ExcelUtils.setCell(x, 1, emp.getLastname().toUpperCase() + ", " + emp.getFirstname().toUpperCase());
+                    ExcelUtils.setCell(x, 2, emp.getStatus().name());
+                    ExcelUtils.setCell(x, 3, emp.getPosition());
+                    x++;
+                }
+                if(detachmentName.length() > 8) {
+                    detachmentName = detachmentName.substring(0, 7);
+                }
+                generateFile(ExcelUtils.workbook, "EMP_"+detachmentName+today.format(DateTimeFormatter.ofPattern("MMddyyyy")));
+                System.out.println("Generation SUCCESS!!");
+                ExcelUtils.workbook.close();
+            } else {
+                addDetailMessage("DETACHMENT HAS NO EMPLOYEES", detachmentName, FacesMessage.SEVERITY_ERROR);
+                FacesContext.getCurrentInstance().getExternalContext().redirect("detachment-form.xhtml?id="+id);
+            }
+        } else {
+            addDetailMessage("NO DETACHMENT", detachmentName, FacesMessage.SEVERITY_ERROR);
+            FacesContext.getCurrentInstance().getExternalContext().redirect("detachment-form.xhtml?id="+id);
+        }
+    }
+
+    public void generateFile(Workbook workbook, String fileName) throws IOException {
+        final FacesContext fc = FacesContext.getCurrentInstance();
+        final ExternalContext externalContext = fc.getExternalContext();
+
+        externalContext.responseReset();
+        externalContext.setResponseContentType("application/vnd.ms-excel");
+        externalContext.setResponseHeader("Content-Disposition", "attachment;filename="+fileName+".xlsx");
+        final HttpServletResponse response = (HttpServletResponse) externalContext.getResponse();
+        final ServletOutputStream out = response.getOutputStream();
+
+        workbook.write(out);
+        workbook.close();
+        out.flush();
+        fc.responseComplete();
+    }
+
+    public static String capitalizeString(String string) {
+        char[] chars = string.toLowerCase().toCharArray();
+        boolean found = false;
+        for (int i = 0; i < chars.length; i++) {
+            if (!found && Character.isLetter(chars[i])) {
+                chars[i] = Character.toUpperCase(chars[i]);
+                found = true;
+            } else if (Character.isWhitespace(chars[i]) || chars[i]=='.' || chars[i]=='\'') { // You can add other chars here
+                found = false;
+            }
+        }
+        return String.valueOf(chars);
     }
 
     public void addDetailMessage(String message, String detail, FacesMessage.Severity severity) {
